Conflicts:
        .travis.yml
	ci/requirements.txt
	qrexec/policy/parser.py
	qrexec/tests/policy_parser.py

+++ daemon/qrexec-daemon.c
@@ -34,6 +34,7 @@
 #include <err.h>
 #include <string.h>
 #include <assert.h>
+#include <getopt.h>
 #include "qrexec.h"
 #include "libqrexec-utils.h"
 
@@ -92,6 +93,10 @@ const char default_user_keyword[] = "DEFAULT:";
 #define default_user_keyword_len_without_colon (sizeof(default_user_keyword)-2)
 
 int opt_quiet = 0;
+int opt_direct = 0;
+
+const char *socket_dir = QREXEC_DAEMON_SOCKET_DIR;
+const char *policy_program = QREXEC_POLICY_PROGRAM;
 
 #ifdef __GNUC__
 #  define UNUSED(x) UNUSED_ ## x __attribute__((__unused__))
@@ -135,9 +140,9 @@ void unlink_qrexec_socket()
     char link_to_socket_name[strlen(remote_domain_name) + sizeof(socket_address)];
 
     snprintf(socket_address, sizeof(socket_address),
-            QREXEC_DAEMON_SOCKET_DIR "/qrexec.%d", remote_domain_id);
+             "%s/qrexec.%d", socket_dir, remote_domain_id);
     snprintf(link_to_socket_name, sizeof link_to_socket_name,
-            QREXEC_DAEMON_SOCKET_DIR "/qrexec.%s", remote_domain_name);
+             "%s/qrexec.%s", socket_dir, remote_domain_name);
     unlink(socket_address);
     unlink(link_to_socket_name);
 }
@@ -155,16 +160,20 @@ int create_qrexec_socket(int domid, const char *domname)
     char link_to_socket_name[strlen(domname) + sizeof(socket_address)];
 
     snprintf(socket_address, sizeof(socket_address),
-            QREXEC_DAEMON_SOCKET_DIR "/qrexec.%d", domid);
+             "%s/qrexec.%d", socket_dir, domid);
     snprintf(link_to_socket_name, sizeof link_to_socket_name,
-            QREXEC_DAEMON_SOCKET_DIR "/qrexec.%s", domname);
+             "%s/qrexec.%s", socket_dir, domname);
     unlink(link_to_socket_name);
+
+    /* When running as root, make the socket accessible; perms on /var/run/qubes still apply */
+    umask(0);
     if (symlink(socket_address, link_to_socket_name)) {
         fprintf(stderr, "symlink(%s,%s) failed: %s\n", socket_address,
                 link_to_socket_name, strerror (errno));
     }
-    atexit(unlink_qrexec_socket);
-    return get_server_socket(socket_address);
+    int fd = get_server_socket(socket_address);
+    umask(0077);
+    return fd;
 }
 
 #define MAX_STARTUP_TIME_DEFAULT 60
@@ -262,53 +271,60 @@ void init(int xid)
             // invalid or negative number
             startup_timeout = MAX_STARTUP_TIME_DEFAULT;
     }
-    signal(SIGUSR1, sigusr1_handler);
-    signal(SIGCHLD, sigchld_parent_handler);
-    switch (pid=fork()) {
-    case -1:
-        perror("fork");
-        exit(1);
-    case 0:
-        break;
-    default:
-        if (getenv("QREXEC_STARTUP_NOWAIT"))
-            exit(0);
-        if (!opt_quiet)
-            fprintf(stderr, "Waiting for VM's qrexec agent.");
-        for (i=0;i<startup_timeout;i++) {
-            sleep(1);
-            if (!opt_quiet)
-                fprintf(stderr, ".");
-            if (i==startup_timeout-1) {
+
+    if (!opt_direct) {
+        signal(SIGUSR1, sigusr1_handler);
+        signal(SIGCHLD, sigchld_parent_handler);
+        switch (pid=fork()) {
+            case -1:
+                perror("fork");
+                exit(1);
+            case 0:
                 break;
-            }
+            default:
+                if (getenv("QREXEC_STARTUP_NOWAIT"))
+                    exit(0);
+                if (!opt_quiet)
+                    fprintf(stderr, "Waiting for VM's qrexec agent.");
+                for (i=0;i<startup_timeout;i++) {
+                    sleep(1);
+                    if (!opt_quiet)
+                        fprintf(stderr, ".");
+                    if (i==startup_timeout-1) {
+                        break;
+                    }
+                }
+                fprintf(stderr, "Cannot connect to '%s' qrexec agent for %d seconds, giving up\n", remote_domain_name, startup_timeout);
+                exit(3);
         }
-        fprintf(stderr, "Cannot connect to '%s' qrexec agent for %d seconds, giving up\n", remote_domain_name, startup_timeout);
-        exit(3);
     }
+
     close(0);
-    snprintf(qrexec_error_log_name, sizeof(qrexec_error_log_name),
-         "/var/log/qubes/qrexec.%s.log", remote_domain_name);
-    umask(0007);        // make the log readable by the "qubes" group
-    logfd =
-        open(qrexec_error_log_name, O_WRONLY | O_CREAT | O_TRUNC,
-         0660);
 
-    if (logfd < 0) {
-        perror("open");
-        exit(1);
-    }
+    if (!opt_direct) {
+        snprintf(qrexec_error_log_name, sizeof(qrexec_error_log_name),
+                 "/var/log/qubes/qrexec.%s.log", remote_domain_name);
+        umask(0007);        // make the log readable by the "qubes" group
+        logfd =
+            open(qrexec_error_log_name, O_WRONLY | O_CREAT | O_TRUNC,
+                 0660);
 
-    dup2(logfd, 1);
-    dup2(logfd, 2);
+        if (logfd < 0) {
+            perror("open");
+            exit(1);
+        }
 
-    if (chdir("/var/run/qubes") < 0) {
-        perror("chdir /var/run/qubes failed");
-        exit(1);
-    }
-    if (setsid() < 0) {
-        perror("setsid()");
-        exit(1);
+        dup2(logfd, 1);
+        dup2(logfd, 2);
+
+        if (chdir("/var/run/qubes") < 0) {
+            perror("chdir /var/run/qubes failed");
+            exit(1);
+        }
+        if (setsid() < 0) {
+            perror("setsid()");
+            exit(1);
+        }
     }
 
     vchan = libvchan_client_init(xid, VCHAN_BASE_PORT);
@@ -339,15 +355,16 @@ void init(int xid)
         vchan_port_notify_client[i] = VCHAN_PORT_UNUSED;
     }
 
-    /* When running as root, make the socket accessible; perms on /var/run/qubes still apply */
-    umask(0);
+    atexit(unlink_qrexec_socket);
     qrexec_daemon_unix_socket_fd =
         create_qrexec_socket(xid, remote_domain_name);
-    umask(0077);
+
     signal(SIGPIPE, SIG_IGN);
     signal(SIGCHLD, sigchld_handler);
     signal(SIGUSR1, SIG_DFL);
-    kill(getppid(), SIGUSR1);   // let the parent know we are ready
+
+    if (!opt_direct)
+        kill(getppid(), SIGUSR1);   // let the parent know we are ready
 }
 
 static int send_client_hello(int fd)
@@ -370,13 +387,39 @@ static int send_client_hello(int fd)
     return 0;
 }
 
-static int allocate_vchan_port(int new_state)
+static int allocate_vchan_port(int connect_domain)
 {
-    int i;
+    /*
+      Make sure the allocated ports numbers are unique for a given {domX, domY}
+      set.
 
-    for (i = 0; i < MAX_CLIENTS; i++) {
+      For domX-domY connections, both daemons can allocate ports. If they both
+      allocate the same port number, this can cause trouble:
+      - We might receive MSG_CONNECTION_TERMINATED for the wrong connection.
+      - Although vchan connections in both directions can exist independently,
+        the direction (client-server or server-client) is not always
+        the same, so collision is still possible.
+
+      To prevent that from happening, for X < Y allow the daemon for X to
+      allocate only odd port numbers, and the daemon for Y to allocate only
+      even port numbers.
+
+      (This does not apply if we are connecting to/from dom0, as there is no
+      separate daemon running for dom0).
+     */
+
+    int i, step;
+    if (connect_domain == 0) {
+        i = 0;
+        step = 1;
+    } else {
+        i = connect_domain > remote_domain_id ? 1 : 0;
+        step = 2;
+    }
+
+    for (; i < MAX_CLIENTS; i += step) {
         if (used_vchan_ports[i] == VCHAN_PORT_UNUSED) {
-            used_vchan_ports[i] = new_state;
+            used_vchan_ports[i] = connect_domain;
             return VCHAN_BASE_DATA_PORT+i;
         }
     }
@@ -797,7 +840,7 @@ static void handle_execute_service(
     signal(SIGPIPE, SIG_DFL);
     snprintf(remote_domain_id_str, sizeof(remote_domain_id_str), "%d",
             remote_domain_id);
-    execl("/usr/bin/qrexec-policy-exec", "qrexec-policy-exec", "--",
+    execl(policy_program, "qrexec-policy-exec", "--",
             remote_domain_id_str,
             remote_domain_name,
             target_domain,
@@ -979,8 +1022,12 @@ static int fill_fdsets_for_select(int vchan_fd, fd_set * read_fdset, fd_set * wr
 /* qrexec-agent has disconnected, cleanup local state and try to connect again.
  * If remote domain dies, terminate qrexec-daemon.
  */
-static int handle_agent_restart(void) {
+static int handle_agent_restart(int xid) {
     size_t i;
+
+    // Stop listening.
+    unlink_qrexec_socket();
+    close(qrexec_daemon_unix_socket_fd);
 
     /* Close old (dead) vchan connection. */
     libvchan_close(vchan);
@@ -1019,7 +1066,33 @@ static int handle_agent_restart(void) {
         return -1;
     }
     fprintf(stderr, "qrexec-agent has reconnected\n");
+
+    qrexec_daemon_unix_socket_fd =
+        create_qrexec_socket(xid, remote_domain_name);
     return 0;
+}
+
+struct option longopts[] = {
+    { "help", no_argument, 0, 'h' },
+    { "quiet", no_argument, 0, 'q' },
+    { "socket-dir", required_argument, 0, 'd' + 128 },
+    { "policy-program", required_argument, 0, 'p' },
+    { "direct", no_argument, 0, 'D' },
+    { NULL, 0, 0, 0 },
+};
+
+_Noreturn void usage(const char *argv0)
+{
+    fprintf(stderr, "usage: %s [options] domainid domain-name [default user]\n", argv0);
+    fprintf(stderr, "Options:\n");
+    fprintf(stderr, "  -h, --help - display usage\n");
+    fprintf(stderr, "  -q, --quiet - quiet mode\n");
+    fprintf(stderr, "  --socket-dir=PATH - directory for qrexec socket, default: %s\n",
+            QREXEC_DAEMON_SOCKET_DIR);
+    fprintf(stderr, "  -p, --policy-program=PATH - program to execute to check policy, default: %s\n",
+            QREXEC_POLICY_PROGRAM);
+    fprintf(stderr, "  -D, --direct - run directly, don't daemonize, log to stderr\n");
+    exit(1);
 }
 
 int main(int argc, char **argv)
@@ -1029,19 +1102,27 @@ int main(int argc, char **argv)
     int max, vchan_fd;
     sigset_t chld_set;
 
-    while ((opt=getopt(argc, argv, "q")) != -1) {
+    while ((opt=getopt_long(argc, argv, "hqp:D", longopts, NULL)) != -1) {
         switch (opt) {
             case 'q':
                 opt_quiet = 1;
                 break;
+            case 'd' + 128:
+                socket_dir = strdup(optarg);
+                break;
+            case 'p':
+                policy_program = strdup(optarg);
+                break;
+            case 'D':
+                opt_direct = 1;
+                break;
+            case 'h':
             default: /* '?' */
-                fprintf(stderr, "usage: %s [-q] domainid domain-name [default user]\n", argv[0]);
-                exit(1);
+                usage(argv[0]);
         }
     }
     if (argc - optind < 2 || argc - optind > 3) {
-        fprintf(stderr, "usage: %s [-q] domainid domain-name [default user]\n", argv[0]);
-        exit(1);
+        usage(argv[0]);
     }
     remote_domain_id = atoi(argv[optind]);
     remote_domain_name = argv[optind+1];
@@ -1083,7 +1164,7 @@ int main(int argc, char **argv)
 
         if (!libvchan_is_open(vchan)) {
             fprintf(stderr, "qrexec-agent has disconnected\n");
-            if (handle_agent_restart() < 0) {
+            if (handle_agent_restart(remote_domain_id) < 0) {
                 fprintf(stderr, "Failed to reconnect to qrexec-agent, terminating\n");
                 return 1;
             }
